+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased changes
 
+## 0.3.0.0
+
 - Feat: collect medium posts by user id
 - Refactor: remove dependent for shelly using threadDelay
 - Refactor: use throwM instead of throwIO
+- Feat: version and verbose options
+- Modify: default no log
 
 ## 0.2.0.0